Revert "Move injected class lookup to class"
This reverts commit 69a12539e6f0606f4bd7b921f1d44658793b4d32.
